+++ BetaProduct/Source/Model/ManagedProduct.m
@@ -10,11 +10,11 @@
 
 @implementation ManagedProduct
 
-+ (instancetype) MakeProductWithName:(NSString *)name WithDescription:(NSString *)description WithPrice:(NSString *)price WithPriceDescription:(NSString *)priceDescription WithLink:(NSString *)link WithImageLink:(NSString *)imageUrl WithCompanyImageLink:(NSString *)companyImageUrl WithProductId:(NSInteger)productId {
++ (instancetype) MakeProductWithName:(NSString *)name WithDescription:(NSString *)description WithPrice:(NSString *)price WithPriceDescription:(NSString *)priceDescription WithLink:(NSString *)link WithImageLink:(NSString *)imageUrl WithCompanyImageLink:(NSString *)companyImageUrl WithProductId:(NSUInteger)productId {
     return [ManagedProduct MakeProductWithName:name WithDescription:description WithPrice:price WithPriceDescription:priceDescription WithLink:link WithImageLink:imageUrl WithCompanyImageLink:companyImageUrl WithProductId:productId WithStatus:StatusActive];
 }
 
-+ (instancetype) MakeProductWithName:(NSString *)name WithDescription:(NSString *)description WithPrice:(NSString *)price WithPriceDescription:(NSString *)priceDescription WithLink:(NSString *)link WithImageLink:(NSString *)imageUrl WithCompanyImageLink:(NSString *)companyImageUrl WithProductId:(NSInteger)productId WithStatus:(Status)status {
++ (instancetype) MakeProductWithName:(NSString *)name WithDescription:(NSString *)description WithPrice:(NSString *)price WithPriceDescription:(NSString *)priceDescription WithLink:(NSString *)link WithImageLink:(NSString *)imageUrl WithCompanyImageLink:(NSString *)companyImageUrl WithProductId:(NSUInteger)productId WithStatus:(Status)status {
     ManagedProduct *product = [[ManagedProduct alloc] init];
     product.name = name;
     product.productDescription = description;
@@ -23,7 +23,7 @@
     product.weblink = link;
     product.imageUrl = imageUrl;
     product.imageCompanyUrl = companyImageUrl;
-    product.productId = (UInt32)productId;
+    product.productId = productId;
     product.status = status;
     return product;
 }
